Make system tests work
We need to have a container just to run the chromedriver and
expose the capybara server port so this container can access the
application.

Using app as the host for the application doesn't work since
chrome automatically make the request HTTPS.

For a future commit we could try to setup a extra_hosts entry in
the app service using the `.test` TLD and use it instead since
it looks better.

Same with the selenium service.