Setup MariaDB container sidecar
This commit adds a MariaDB container to the docker-compose.yml file and
configures Rails to use it.

There are a few things to note:

* Rails by default connects to the root user with no password in
development and test. This is not a good practice, but since it is
what we already do, I set up the container to match that with the
`MARIADB_ALLOW_EMPTY_ROOT_PASSWORD` environment variable.

* Rails by default connects to localhost. But we aren't running the
container in the host network, so we need to use the container name
in the `database.yml` file.